NEX-39: -editing page

+++ Next2/src/Next2/Models/TaxModel.cs
@@ -10,6 +10,6 @@ namespace Next2.Models
     {
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public double Value { get; set; }
+        public float Value { get; set; }
     }
 }
